Hook up real Inbox panel content.
